Update gql, typedefs, and resolver to return forwarded users in items responses
