+++ api/typeDefs/itemForward.js
@@ -1,0 +1,13 @@
+import { gql } from 'graphql-tag'
+
+export default gql`
+  type ItemForward {
+    id: ID!
+    created_at: Date!
+    updated_at: Date!
+    itemId: Int!
+    userId: Int!
+    user: User!
+    pct: Int!
+  }
+`
