+++ fragments/items.js
@@ -43,6 +43,13 @@ export const ITEM_FIELDS = gql`
     status
     uploadId
     mine
+    forwards {
+      userId
+      pct
+      user {
+        name
+      }
+    }
   }`
 
 export const ITEM_FULL_FIELDS = gql`
